requirements.txt
flask
gunicorn
matplotlib
pytest
python-dotenv
flask-session
